requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.10
django-debug-toolbar==2.2
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
